# app/Http/Requests/UpdateIncidentStatusRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;
use Illuminate\Validation\Rule;

class UpdateIncidentStatusRequest extends FormRequest
{
    public function authorize(): bool
    {
        return true;
    }

    public function rules(): array
    {
        return [
            'status' => ['required', 'string', Rule::in(array_keys(config('incidents.statuses', [])))],
            'admin_notes' => ['nullable', 'string', 'max:2000'],
        ];
    }
}