# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Postos API",
        "description": "API de gestão de postos (PinPoint): JWT, RBAC, stock e histórico.\r\n\r\n**Rotas públicas (sem JWT):**\r\n- `GET /postos` — listar, paginar (`?page=`) ou detalhe (`?id=`)\r\n- `GET /postos/search` — pesquisa rápida (`?q=`, opcional `limit`, `tipo`, `combustivel`, `status`)\r\n\r\n**Papéis locais (`users.role_id`):** ver `docs/ROLES.md` e `GET /api/roles`.",
        "contact": {
            "email": "suporte@postosapi.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Prefixo das rotas API"
        }
    ],
    "paths": {
        "/admin/incidents": {
            "get": {
                "tags": [
                    "Incidentes"
                ],
                "summary": "Listar todos os incidentes (admin)",
                "description": "RN-G-006: visão global para resolução. Filtros: status, category, post_id.",
                "operationId": "93855f23750ee5cc059ffed8b092af6d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "urgente",
                                "normal"
                            ]
                        }
                    },
                    {
                        "name": "post_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada"
                    },
                    "403": {
                        "description": "Apenas admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/incidents/{incident}": {
            "patch": {
                "tags": [
                    "Incidentes"
                ],
                "summary": "Actualizar estado do incidente (admin)",
                "operationId": "3157a979d661f08a742f8183412e6fd5",
                "parameters": [
                    {
                        "name": "incident",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIncidentStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Apenas admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/price-decrees": {
            "get": {
                "tags": [
                    "Preços"
                ],
                "summary": "Listar preços decretados (admin)",
                "operationId": "0983854ef95cc2222b7721aa465a57bb",
                "responses": {
                    "200": {
                        "description": "Lista de decretos"
                    },
                    "403": {
                        "description": "Apenas admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Preços"
                ],
                "summary": "Publicar preço decretado ANPG/IRDP (admin)",
                "description": "RN-G-001: define preços; gestores apenas confirmam aplicação no posto.",
                "operationId": "c9533727fec21867deea42f3be6ee3e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublishPriceDecreeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Decreto publicado"
                    },
                    "403": {
                        "description": "Apenas admin"
                    },
                    "422": {
                        "description": "Validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/campaigns/nearby": {
            "get": {
                "tags": [
                    "Campanhas Públicas"
                ],
                "summary": "Campanhas activas perto do utilizador (RN-G-007)",
                "description": "Rota pública. Valida proximidade geográfica ao posto (raio da campanha).",
                "operationId": "9c388c0628729242b885ec5896157cd1",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "post_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campanhas próximas"
                    }
                }
            }
        },
        "/campaigns/{campaign}/interactions": {
            "post": {
                "tags": [
                    "Campanhas Públicas"
                ],
                "summary": "Registar visualização/clique/conversão",
                "description": "Exige coordenadas dentro do raio geográfico da campanha.",
                "operationId": "e257c1c363d391d155b165f09df52392",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TrackCampaignInteractionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interacção registada"
                    },
                    "422": {
                        "description": "Fora do raio ou campanha inactiva"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PublishPriceDecreeRequest": {
                "required": [
                    "preco"
                ],
                "properties": {
                    "reference": {
                        "type": "string",
                        "example": "ANPG/IRDP Março 2026"
                    },
                    "fuel_type_id": {
                        "description": "1=Gasolina, 2=Gasóleo, etc.",
                        "type": "integer",
                        "example": 1
                    },
                    "preco": {
                        "type": "string",
                        "example": "320 Kz/L"
                    },
                    "preco_premium": {
                        "type": "string",
                        "example": "340 Kz/L"
                    },
                    "effective_from": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "confirmation_deadline_hours": {
                        "type": "integer",
                        "example": 48
                    }
                },
                "type": "object"
            },
...
            "UpdateIncidentStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "aberto",
                            "em_andamento",
                            "resolvido",
                            "cancelado"
                        ]
                    },
                    "admin_notes": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "TrackCampaignInteractionRequest": {
                "required": [
                    "event_type",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "event_type": {
                        "type": "string",
                        "enum": [
                            "view",
                            "click",
                            "conversion"
                        ]
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "client_user_id": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "IncidentResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "post_id": {
                        "type": "integer"
                    },
                    "category": {
                        "type": "string"
                    },
                    "equipment_type": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "photos": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT no header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}